backend/requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Image processing
pillow>=10.0.0
pytesseract>=0.3.10  # For OCR fallback

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API clients
anthropic>=0.18.0
openai>=1.12.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Monitoring
prometheus-client>=0.19.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0

# Evaluation
pytest>=7.4.0
scikit-learn>=1.3.0

# Optional: UI
streamlit>=1.30.0  # For web interface

# Optional: Advanced vision
transformers>=4.36.0  # For CLIP/Florence-2
torch>=2.1.0  # For vision models

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.8.0

# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6  # For file uploads

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0  # PostgreSQL async driver
alembic>=1.13.0  # Database migrations

# Production ASGI server
gunicorn>=21.2.0

# Additional utilities
pydantic>=2.5.0
python-json-logger>=2.0.7  # Structured logging

# Optional: Performance
orjson>=3.9.0  # Faster JSON
httptools>=0.6.0  # Faster HTTP parsing

# Optional: Monitoring
prometheus-client>=0.19.0  # Metrics
sentry-sdk>=1.39.0  # Error tracking

# Optional: Rate limiting
slowapi>=0.1.9

# Optional: Authentication
python-jose[cryptography]>=3.3.0  # JWT
passlib[bcrypt]>=1.7.4  # Password hashing

# Testing
httpx>=0.26.0  # Async HTTP client for tests
pytest-asyncio>=0.23.0

prometheus-fastapi-instrumentator
